ALX SE

+++ 0x17-doubly_linked_lists/8-delete_dnodeint.c
@@ -30,7 +30,7 @@ int delete_dnodeint_at_index(dlistint_t **head, unsigned int index)
 		}
 		*head = (*head)->next;
 	}
-	if ((int)index > i)
+	if (index > i)
 		return (-1);
 	return (1);
 }
